chore: revert auto format changes

+++ src/interfaces/AccountStructs.sol
@@ -5,85 +5,85 @@ import "./IAsset.sol";
 
 // For full documentation refer to src/Account.sol";
 interface AccountStructs {
-    /////////////////////
-    // Storage Structs //
-    /////////////////////
+  /////////////////////
+  // Storage Structs //
+  /////////////////////
 
-    struct BalanceAndOrder {
+  struct BalanceAndOrder {
     // balance of (asset, subId)
-        int240 balance;
-        // index in heldAssets() or getAccountBalances()
-        uint16 order;
-    }
+    int240 balance;
+    // index in heldAssets() or getAccountBalances()
+    uint16 order;
+  }
 
-    struct HeldAsset {
-        IAsset asset;
-        uint96 subId;
-    }
+  struct HeldAsset {
+    IAsset asset;
+    uint96 subId;
+  }
 
-    struct AssetDelta {
-        IAsset asset;
-        uint96 subId;
-        int256 delta;
-    }
+  struct AssetDelta {
+    IAsset asset;
+    uint96 subId;
+    int delta;
+  }
 
-    // the struct is used to easily manage 2 dimensional array
-    struct AssetDeltaArrayCache {
-        uint256 used;
-        AssetDelta[100] deltas;
-    }
+  // the struct is used to easily manage 2 dimensional array
+  struct AssetDeltaArrayCache {
+    uint used;
+    AssetDelta[100] deltas;
+  }
 
-    /////////////////////////
-    // Memory-only Structs //
-    /////////////////////////
+  /////////////////////////
+  // Memory-only Structs //
+  /////////////////////////
 
-    struct AssetBalance {
-        IAsset asset;
-        // adjustments will revert if > uint96
-        uint256 subId;
-        // base layer only stores up to int240
-        int256 balance;
-    }
+  struct AssetBalance {
+    IAsset asset;
+    // adjustments will revert if > uint96
+    uint subId;
+    // base layer only stores up to int240
+    int balance;
+  }
 
-    struct AssetTransfer {
+  struct AssetTransfer {
     // credited by amount
-        uint256 fromAcc;
-        // debited by amount
-        uint256 toAcc;
-        IAsset asset;
-        // adjustments will revert if >uint96
-        uint256 subId;
-        // reverts if transfer amount > uint240
-        int256 amount;
-        // data passed into asset.handleAdjustment()
-        bytes32 assetData;
-    }
+    uint fromAcc;
+    // debited by amount
+    uint toAcc;
+    IAsset asset;
+    // adjustments will revert if >uint96
+    uint subId;
+    // reverts if transfer amount > uint240
+    int amount;
+    // data passed into asset.handleAdjustment()
+    bytes32 assetData;
+  }
 
-    struct AssetAdjustment {
-        uint256 acc;
-        IAsset asset;
-        // reverts for subIds > uint96
-        uint256 subId;
-        // reverts if transfer amount > uint240
-        int256 amount;
-        // data passed into asset.handleAdjustment()
-        bytes32 assetData;
-    }
+  struct AssetAdjustment {
+    uint acc;
+    IAsset asset;
+    // reverts for subIds > uint96
+    uint subId;
+    // reverts if transfer amount > uint240
+    int amount;
+    // data passed into asset.handleAdjustment()
+    bytes32 assetData;
+  }
 
-    ////////////////
-    // Allowances //
-    ////////////////
+  ////////////////
+  // Allowances //
+  ////////////////
 
-    struct AssetAllowance {
-        IAsset asset;
-        uint256 positive;
-        uint256 negative;
-    }
+  struct AssetAllowance {
+    IAsset asset;
+    uint positive;
+    uint negative;
+  }
 
-    struct SubIdAllowance {
-        IAsset asset;
-        uint256 subId;
-        uint256 positive;
-        uint256 negative;
-    }
+  struct SubIdAllowance {
+    IAsset asset;
+    uint subId;
+    uint positive;
+    uint negative;
+  }
 }
