+++ test/account/mocks/feeds/IVFeeds.sol
@@ -6,32 +6,26 @@ import "src/interfaces/IAsset.sol";
 
 // simple single IV oracle
 contract IVFeeds is Owned {
-    mapping(IAsset => mapping(uint256 => uint256)) public feedIds; // asset => subId => feedId;
-    mapping(uint256 => uint256) IVs; // feedId => iv;
+  mapping(IAsset => mapping(uint => uint)) public feedIds; // asset => subId => feedId;
+  mapping(uint => uint) IVs; // feedId => iv;
 
-    constructor() Owned() {}
+  constructor() Owned() {}
 
-    function setIVForFeed(uint256 feedId, uint256 iv) external onlyOwner {
-        IVs[feedId] = iv;
-    }
+  function setIVForFeed(uint feedId, uint iv) external onlyOwner {
+    IVs[feedId] = iv;
+  }
 
-    function getIVForFeed(uint256 feedId) public view returns (uint256 iv) {
-        iv = IVs[feedId];
-        require(iv != 0, "No iv set for asset, subId");
-        return iv;
-    }
+  function getIVForFeed(uint feedId) public view returns (uint iv) {
+    iv = IVs[feedId];
+    require(iv != 0, "No iv set for asset, subId");
+    return iv;
+  }
 
-    function assignFeedToSubId(IAsset asset, uint256 subId, uint256 feedId)
-        external
-    {
-        feedIds[asset][subId] = feedId;
-    }
+  function assignFeedToSubId(IAsset asset, uint subId, uint feedId) external {
+    feedIds[asset][subId] = feedId;
+  }
 
-    function getIVForSubId(IAsset asset, uint256 subId)
-        external
-        view
-        returns (uint256 iv)
-    {
-        return getIVForFeed(feedIds[asset][subId]);
-    }
+  function getIVForSubId(IAsset asset, uint subId) external view returns (uint iv) {
+    return getIVForFeed(feedIds[asset][subId]);
+  }
 }
